refactor: PlayerController로 병합
